Fixing exception catching in process....2

+++ src/main/java/edu/epam/project/service/AdminService.java
@@ -3,7 +3,7 @@ package edu.epam.project.service;
 import edu.epam.project.entity.User;
 import edu.epam.project.exception.ServiceException;
 
-public interface AdminService extends BaseService<Long, User>{
+public interface AdminService extends BaseService<Integer, User>{
 
     boolean activateHR(User user) throws ServiceException;
